move sub line trimming into sub parser
It occured to me that if we ever want to use this for something else
like TL notes that might use a /different/ parser, this could cause me a
headache trying to figure out why the first character is skipped because
I put it way too late in the process. Plus, this is two more operations
that don't need to happen on every frame so double win.