move hook to make triggers correct and misc fix
Fixed by changing from a function that executes /before/ the line is set
to one that actually sets the line. This means that triggers will
execute almost immediately and more importantly - aligned with the line
they're expected to execute on.

Changed all debug string allocations to be on the stack