add retrying to d3d9 hook and rework ending tracks
Before, the only thing that set the playing flag to false was hitting
the end trigger, but I have a feeling this is causing some of the
crashes Krizs is seeing.

The downside of this is that we're now displaying ticks 24/7 on the
debug menu. I could add another flag to see if this should stop, but I
don't want to do that unless its driving people crazy.