add triggers for voice lines
Because reasons, Leaf was not particularly consistent in their chosen
ways to initiate audio playback during these no-dialogue sections. For
that reason, there's now at least /two/ functions we need to watch. The
new one added here is triggered when every voice line is played. As a
result of that, there needs to be a new triggering system.