add back line-load hook and refactor hooks
Line loading is necessary for hard-cutoffs in any case when there's no
audio track being loaded on a line (every Haruki monologue line). This
also removes a check from the linear search of all lines meaning we
technically have slightly lower coverage of cases because of the
following:

1. Sub track starts playing
2. User loads a save mid-track in the same file
3. the sub track continues playing because theres no cutoff
4. User hits a hard-cutoff for an entirely different track

This won't work anymore but... this is a pretty stupid edge-case and if
I really cared about this kinda thing I'd add proper logic to handle
this.