adjust text top adjust for 'd', 'l', and 't'
The font is always gonna be a little bit wrong just because it's a
monospace renderer using a variable-width set of glyphs. With this it's
going to look consistently /ok/ which I think is more than enough. I
just don't want it to look distractingly bad which the characters I
fixed here did.