redesign d3d9 hook
Originally, this hook worked simply by waiting and then trying to grab
the existing d3d9 context. However, this had a timing issue that
appeared on a bunch of systems when we did our first round of testing.
As such, I decided to wrap the direct3d objects so that theres no race
condition going on.