add skeleton for DirectShow stub
This is the first step in trying to swap out the video player for the
engine. For unknown reasons, the engine will sometimes just play a black
video and the DirectShow backend doesn't work for WINE users.

I'm going to try swapping out the video player backend for something a
little more cross-platform friendly like ffmpeg.

Right now this hook does nothing but call into the original function and
stubbing it out doesn't really work because it just hangs on startup so
I'm first gonna see if I can fix whatever bug is causing black screens
on windows. If I can get that then I might just be satisfied. I might
also try to swap out the player for the DirectShow MP4 decoding instead
of WMV.